backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
google-auth
